+++ util.h
@@ -6,6 +6,6 @@
 #define STRLEN 100
 
 pj_str_t QstrToPstr(QString src,char dest[]);
-pj_str_t QstrToPstr(QString src,char dest[],size_t size);
+pj_str_t QstrToPstr(QString src,char dest[],int size);
 pj_str_t QstrToPstr(QString src);
 #endif // UTIL_H
